removed old bluetooth functionality
