+++ PlaygroundBook/Sources/CalliopeImage.swift
@@ -1,3 +1,23 @@
+public enum CalliopeImage: Int16 {
+	case smiley = 0x00
+	case sad = 0x01
+	case heart = 0x02
+	case arrow_left = 0x03
+	case arrow_right = 0x04
+	case arrow_left_right = 0x05
+	case full = 0x06
+	case dot = 0x07
+	case small_rect = 0x08
+	case large_rect = 0x09
+	case double_row = 0x0a
+	case tick = 0x0b
+	case rock = 0x0c
+	case scissors = 0x0d
+	case well = 0x0e
+	case flash = 0x0f
+	case wave = 0x10
+}
+
 extension CalliopeImage: RawRepresentable {
     
     static var all: [CalliopeImage] {
@@ -15,5 +35,4 @@ extension CalliopeImage: RawRepresentable {
         guard let type = CalliopeImage.all.first(where: { "."+String(describing:$0) == from }) else { return nil }
         self = type
     }
-    
 }
